# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PocketSync API
  description: |
    PocketSync REST API.

    Production: `https://pocketsync.onrender.com`

    Auth uses HttpOnly cookies (`accessToken`, `refreshToken`). Send cookies on protected
    routes (`credentials: 'include'` / `withCredentials: true`).

    Sign up: register → verify-otp → login.
    Password reset: forgot-password → reset-password.
    BVN onboarding: status → bvn/submit → bvn/verify-otp → bvn/accounts → bvn/connect.

    BVN, SMS, external transfers, and bill pay are simulated in this build.
    OTPs log to the server console; `devOtp` may appear in dev responses.

    Amounts are in NGN. Credits are positive; debits are negative in transaction objects.
  version: 1.0.0
  contact:
    name: PocketSync Backend Team
servers:
  - url: https://pocketsync.onrender.com
    description: Production (Render)
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/v1/transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      operationId: getTransactions
      security:
        - cookieAuth: []
      parameters:
        - name: accountId
          in: query
          schema:
            type: string
          description: Filter by linked account ID (must belong to user)
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/TransactionCategory'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          example: "2026-01-01"
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          example: "2026-06-26"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            maximum: 30
      responses:
        '200':
          description: Paginated transaction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Account not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/transactions/transfer:
    post:
      tags: [Transactions]
      summary: Internal transfer (between your linked accounts)
      description: |
        Moves money between two of the user's own linked accounts.
        Creates paired debit + credit transactions with the same reference.
      operationId: transferMoney
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalTransferRequest'
            example:
              fromAccountId: "6a3e576b28a32290927300a5"
              toAccountId: "6a3e576b28a32290927300b6"
              amount: 5000
              description: Transfer to Access Bank savings
      responses:
        '201':
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalTransferResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Account not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/transactions/interbank-transfer:
    post:
      tags: [Transactions]
      summary: Interbank transfer (to external account)
      description: |
        External bank transfer (simulated NIP).
        `recipientBank` accepts any valid bank name.
        Debits the source account only.
      operationId: interbankTransfer
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterbankTransferRequest'
            example:
              fromAccountId: "6a3e576b28a32290927300c8"
              recipientBank: UBA
              recipientAccountNumber: "0123456789"
              recipientName: khalifa bin zayed
              amount: 10000
              description: Payment for services
      responses:
        '201':
          description: Interbank transfer successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterbankTransferResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/transactions/pay-bill:
    post:
      tags: [Transactions]
      summary: Pay a bill
      operationId: payBill
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayBillRequest'
            example:
              fromAccountId: "6a3e576b28a32290927300a5"
              amount: 21500
              billProvider: DSTV
              customerReference: "1234567890"
              description: DSTV Premium renewal
      responses:
        '201':
          description: Bill payment successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayBillResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/transactions/{transactionId}:
    get:
      tags: [Transactions]
      summary: Get single transaction
      operationId: getTransactionById
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/transactionId'
      responses:
        '200':
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HttpOnly cookie set by POST /api/v1/auth/login
  parameters:
...
    transactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectId of transaction
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorised — please log in
    Forbidden:
      description: Resource exists but not owned by user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          description: Present in development only
...
    LinkableInstitution:
      type: string
      enum: [GTBank, Access Bank, Kuda, Opay, Moniepoint]
    AccountType:
      type: string
      enum: [current, savings, wallet, business]
    TransactionCategory:
      type: string
      enum: [Food, Transport, Bills, Entertainment, Savings, Transfer, Other]
    TransactionType:
      type: string
      enum: [credit, debit]
    BillProvider:
      type: string
      enum: [DSTV, GOTV, IKEDC, EKEDC, MTN, Airtel, Glo, 9mobile, LAWMA, Water Board]
    LinkedAccount:
      type: object
      properties:
        id:
          type: string
        institution:
          $ref: '#/components/schemas/LinkableInstitution'
        maskedAccountNumber:
          type: string
          example: "****4471"
        balance:
          type: number
          description: Balance in NGN
        currency:
          type: string
          example: NGN
        accountType:
          $ref: '#/components/schemas/AccountType'
        linkedAt:
          type: string
          format: date-time
    Transaction:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
          format: date-time
        description:
          type: string
        amount:
          type: number
          description: NGN. Negative for debits, positive for credits.
        type:
          $ref: '#/components/schemas/TransactionType'
        category:
          $ref: '#/components/schemas/TransactionCategory'
        institution:
          type: string
        accountId:
          type: string
        reference:
          type: string
...
    TransactionsListResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    InternalTransferRequest:
      type: object
      required: [fromAccountId, toAccountId, amount]
      properties:
        fromAccountId:
          type: string
        toAccountId:
          type: string
        amount:
          type: number
          minimum: 1
          description: Amount in NGN
        description:
          type: string
    InternalTransferResponse:
      type: object
      properties:
        message:
          type: string
        transfer:
          type: object
          properties:
            reference:
              type: string
            amount:
              type: number
            fromAccount:
              $ref: '#/components/schemas/LinkedAccount'
            toAccount:
              $ref: '#/components/schemas/LinkedAccount'
            debitTransaction:
              $ref: '#/components/schemas/Transaction'
            creditTransaction:
              $ref: '#/components/schemas/Transaction'
    InterbankTransferRequest:
      type: object
      required: [fromAccountId, recipientBank, recipientAccountNumber, amount]
      properties:
        fromAccountId:
          type: string
        recipientBank:
          type: string
          minLength: 2
          maxLength: 100
          description: Any Nigerian bank name (e.g. UBA, Zenith Bank, First Bank)
        recipientAccountNumber:
          type: string
          pattern: '^\\d{10}$'
        recipientName:
          type: string
        amount:
          type: number
          minimum: 1
        description:
          type: string
    InterbankTransferResponse:
      type: object
      properties:
        message:
          type: string
        transfer:
          type: object
          properties:
            reference:
              type: string
            nipReference:
              type: string
            amount:
              type: number
            status:
              type: string
              example: completed
            fromAccount:
              $ref: '#/components/schemas/LinkedAccount'
            recipient:
              type: object
              properties:
                bank:
                  type: string
                accountNumber:
                  type: string
                  example: "****6789"
                name:
                  type: string
            transaction:
              $ref: '#/components/schemas/Transaction'
    PayBillRequest:
      type: object
      required: [fromAccountId, amount, billProvider]
      properties:
        fromAccountId:
          type: string
        amount:
          type: number
          minimum: 1
        billProvider:
          $ref: '#/components/schemas/BillProvider'
        customerReference:
          type: string
          description: Meter number, smartcard ID, phone number, etc.
        description:
          type: string
    PayBillResponse:
      type: object
      properties:
        message:
          type: string
        payment:
          type: object
          properties:
            reference:
              type: string
            amount:
              type: number
            billProvider:
              $ref: '#/components/schemas/BillProvider'
            account:
              $ref: '#/components/schemas/LinkedAccount'
            transaction:
              $ref: '#/components/schemas/Transaction'
...